Claude's completion of the lecture
Interestingly, Claude needed prompting to find the deliberate mistake in
GLGetters.hpp. In previous rehearsals it noticed the mistake. That may
have been because this time I opted to neglect the static_cast. Once
alerted to the problem, Claude did indeed fix it and put the static_cast
in itself, without additional prompting. I'm not sure, however, if
Claude realized that the static_cast isn't strictly necessary; it is,
however, good practice to highlight the conversion.

+++ Source/avocet/OpenGL/Context/GLGetters.hpp
@@ -381,7 +381,9 @@ namespace avocet::opengl {
     template<std::derived_from<context_base> Context>
     [[nodiscard]]
     GLuint get(const Context& ctx, mask_names name) {
-        return impl::do_get<GLuint>(ctx, name);
+        // Masks have no dedicated unsigned getter: GetIntegerv is the only way to query them.
+        // We therefore read through the GLint path and reinterpret the bit pattern as GLuint.
+        return static_cast<GLuint>(impl::do_get<GLint>(ctx, name));
     }
 
     template<std::derived_from<context_base> Context>
